requirements.txt
psycopg2-binary
pgvector
requests
kafka-python
python-dotenv
beautifulsoup4
sqlalchemy
openai
feedparser